+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Some file dialogs would not remember the last used directory.
  - Saving a ROM when no SAV has been loaded.
 
+### Added
+ - Kit Editor: "Reload samples" button.
+
 ### Changed
  - Kit Editor: "Add Sample" now automatically resamples, normalizes and dithers the sample. No need to prepare samples using sox anymore.
  - Kit Editor: switched to TPDF dithering noise, which gives way better audio quality than the old one.